IOS-3828 Add changes after review comments
